refactor: Компонент favorites-screen разбит на несколько файлов
