+++ src/components/app/app.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MainScreen from '../main-screen/main-screen';
 import NotFoundScreen from '../not-found-screen/not-found-screen';
 import LoginScreen from '../login-screen/login-screen';
-import FavoritesScreen from '../favorites-screen/favorites-screen';
+import FavoritesScreen from '../favorites/favorites-screen';
 import PropertyScreen from '../property-screen/property-screen';
 
 import PropTypes from 'prop-types';
